+++ extensions/mozilla-firefox/src/new-tab.tsx
@@ -1,21 +1,15 @@
-import { useState, ReactElement } from "react";
+import { ReactElement } from "react";
 import { List } from "@raycast/api";
 import { useHistorySearch } from "./hooks/useHistorySearch";
+import { useEditUrlInSearch } from "./hooks/useEditUrlInSearch";
 import { HistoryListEntry, NewTabEntry } from "./components";
 
 export default function Command(): ReactElement {
-  const [searchText, setSearchText] = useState("");
-  const [selectedItemId, setSelectedItemId] = useState<string | undefined>();
+  const { searchText, setSearchText, selectedItemId, editUrlInSearch } = useEditUrlInSearch();
   const { isLoading: isLoadingHistory, errorView: errorHistory, data: entriesHistory } = useHistorySearch(searchText);
 
   if (errorHistory) {
     return errorHistory;
-  }
-
-  function editUrlInSearch(url: string) {
-    setSearchText(url);
-    setSelectedItemId("new-tab");
-    setTimeout(() => setSelectedItemId(undefined), 0);
   }
 
   return (
@@ -24,7 +18,7 @@ export default function Command(): ReactElement {
       onSearchTextChange={setSearchText}
       selectedItemId={selectedItemId}
       isLoading={isLoadingHistory}
-      throttle={false}
+      throttle={true}
     >
       <List.Section title="New Tab" key="new-tab">
         <NewTabEntry searchText={searchText} />