+++ extensions/mozilla-firefox/src/search-bookmarks.tsx
@@ -1,21 +1,15 @@
 import { List } from "@raycast/api";
-import { useState, ReactElement } from "react";
+import { ReactElement } from "react";
 import { HistoryListEntry, NewTabEntry } from "./components";
 import { useBookmarkSearch } from "./hooks/useBookmarkSearch";
+import { useEditUrlInSearch } from "./hooks/useEditUrlInSearch";
 
 export default function Command(): ReactElement {
-  const [searchText, setSearchText] = useState("");
-  const [selectedItemId, setSelectedItemId] = useState<string | undefined>();
+  const { searchText, setSearchText, selectedItemId, editUrlInSearch } = useEditUrlInSearch();
   const { isLoading, errorView, data } = useBookmarkSearch(searchText);
 
   if (errorView) {
     return errorView;
-  }
-
-  function editUrlInSearch(url: string) {
-    setSearchText(url);
-    setSelectedItemId("new-tab");
-    setTimeout(() => setSelectedItemId(undefined), 0);
   }
 
   return (
@@ -24,7 +18,7 @@ export default function Command(): ReactElement {
       onSearchTextChange={setSearchText}
       selectedItemId={selectedItemId}
       isLoading={isLoading}
-      throttle={false}
+      throttle={true}
     >
       <List.Section title="Open URL" key="open-url">
         <NewTabEntry searchText={searchText} />