+++ extensions/mozilla-firefox/src/search-history.tsx
@@ -1,7 +1,8 @@
 import { List } from "@raycast/api";
 import { useHistorySearch } from "./hooks/useHistorySearch";
-import { useState, ReactElement } from "react";
+import { ReactElement } from "react";
 import { HistoryListEntry, NewTabEntry } from "./components";
+import { useEditUrlInSearch } from "./hooks/useEditUrlInSearch";
 import { GroupedEntries, HistoryEntry } from "./interfaces";
 
 const groupEntries = (allEntries?: HistoryEntry[]): GroupedEntries =>
@@ -21,18 +22,11 @@ const groupEntries = (allEntries?: HistoryEntry[]): GroupedEntries =>
     : new Map<string, HistoryEntry[]>();
 
 export default function Command(): ReactElement {
-  const [searchText, setSearchText] = useState("");
-  const [selectedItemId, setSelectedItemId] = useState<string | undefined>();
+  const { searchText, setSearchText, selectedItemId, editUrlInSearch } = useEditUrlInSearch();
   const { isLoading, errorView, data } = useHistorySearch(searchText);
 
   if (errorView) {
     return errorView;
-  }
-
-  function editUrlInSearch(url: string) {
-    setSearchText(url);
-    setSelectedItemId("new-tab");
-    setTimeout(() => setSelectedItemId(undefined), 0);
   }
 
   const groupedEntries = groupEntries(data);